refactor: refactor into a library to be referenced by cli tools and actions
