+++ __tests__/common.test.ts
@@ -1,0 +1,13 @@
+
+
+import { describe, expect, test } from '@jest/globals';
+import { parseRepositories } from '../src/common';
+
+describe("input parsing", () => {
+    test("parse repositories input", () => {
+      const repos = parseRepositories("owner/repo1,owner/repo2")
+      expect(repos[0]).toStrictEqual({ "repo": "repo1", "owner": "owner" })
+      expect(repos[1]).toStrictEqual({ "repo": "repo2", "owner": "owner" })
+    });
+  });
+
