--- label-push/action.yml
@@ -1,12 +1,0 @@
-name: 'label-push'
-description: 'Distribute the labals from a source repository to other repositories.'
-inputs:
-  repositories:
-    description: 'Comma delimited list of repositories to synchronise.'
-    required: true
-  token:
-    description: 'Token with access to repositories you wish to push to.'
-    required: true
-runs:
-  using: 'node16'
-  main: '../dist/label-push/index.js'
