+++ package.json
@@ -1,49 +1,58 @@
 {
-  "name": "multirepo-project-actions",
-  "version": "0.0.0",
+  "name": "mr-tools-lib",
+  "version": "0.0.0-development",
   "private": true,
   "license": "MIT",
+  "type": "module",
   "scripts": {
     "build": "tsc --build",
-    "package:milestone-push": "ncc build ./lib/milestone-push.js -o ./dist/milestone-push",
-    "package:label-push": "ncc build ./lib/label-push.js -o ./dist/label-push",
-    "package": "npm-run-all package:*",
     "test": "jest",
     "lint": "eslint --report-unused-disable-directives --max-warnings=0 . --ext .ts",
-    "lint-fix": "eslint --report-unused-disable-directives --max-warnings=0 . --ext .ts --fix",
-    "all": "npm-run-all build test package",
-    "clean": "rimraf ./dist ./lib ./coverage ./node_modules"
+    "clean": "rimraf ./dist ./lib ./coverage ./node_modules",
+    "semantic-release": "semantic-release"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/deanhigh/multirepo-project-actions.git"
+    "url": "https://github.com/deanhigh/mr-tools-lib.git"
   },
   "bugs": {
-    "url": "https://github.com/deanhigh/multirepo-project-actions/issues"
+    "url": "https://github.com/deanhigh/mr-tools-lib/issues"
   },
-  "homepage": "https://github.com/deanhigh/multirepo-project-actions#readme",
+  "homepage": "https://github.com/deanhigh/mr-tools-lib#readme",
   "devDependencies": {
-    "@jest/globals": "^29.3.1",
+    "@jest/globals": "^29.4.1",
+    "@semantic-release/changelog": "^6.0.2",
+    "@semantic-release/git": "^10.0.1",
+    "@semantic-release/npm": "^9.0.2",
+    "@semantic-release/release-notes-generator": "^10.0.3",
     "@types/js-yaml": "^4.0.5",
     "@types/node": "^18.11.18",
-    "@typescript-eslint/eslint-plugin": "^5.48.1",
-    "@vercel/ncc": "^0.36.0",
-    "eslint": "^8.31.0",
+    "@typescript-eslint/eslint-plugin": "^5.49.0",
+    "eslint": "^8.33.0",
     "eslint-plugin-github": "^4.6.0",
+    "eslint-plugin-jest": "^25.3.2",
     "eslint-plugin-no-async-foreach": "^0.1.1",
-    "jest": "^29.3.1",
-    "npm-run-all": "^4.1.5",
-    "rimraf": "^4.0.4",
+    "jest": "^29.4.1",
+    "prettier": "2.8.3",
+    "rimraf": "^4.1.2",
+    "semantic-release": "^20.1.0",
+    "ts-jest": "^29.0.5",
     "typescript": "^4.9.4"
   },
   "dependencies": {
     "@actions/core": "^1.10.0",
     "@actions/github": "^5.1.1",
     "js-yaml": "^4.1.0",
-    "ts-jest": "^29.0.3",
     "ts-node": "^10.9.1"
   },
   "release": {
+    "plugins": [
+      "@semantic-release/release-notes-generator",
+      "@semantic-release/changelog",
+      "@semantic-release/npm",
+      "@semantic-release/git",
+      "@semantic-release/github"
+    ],
     "branches": [
       "main",
       "next"
