--- src/label-push.ts
@@ -1,3 +1,0 @@
-// import {milestonePush} from "./common"
-
-console.log("TO BT IMPLEMENTED");
