--- src/milestone-push.ts
@@ -1,3 +1,0 @@
-import { milestonePush } from "./common";
-
-void milestonePush();
